implement TestCommon
Extract auxiliary functions from TestSingleGradient to TestCommon.
